Move casting to graph object to object itself
